+++ console/console-init/ui/src/graphql-module/queries/Address.ts
@@ -154,6 +154,9 @@ const RETURN_ALL_ADDRESS_FOR_ADDRESS_SPACE = (
           spec {
             address
             type
+            topic
+            deadletter
+            expiry
             plan {
               spec {
                 displayName
@@ -235,7 +238,9 @@ const RETURN_ADDRESS_DETAIL = (
             address
             type
             topic
-            plan {
+            deadletter
+            expiry
+              plan {
               spec {
                 displayName
                 addressType
@@ -489,6 +494,50 @@ const RETURN_ADDRESS_SPACE_PLANS = gql`
   }
 `;
 
+const RETURN_DLQ_ADDRESSES_FOR_SUBSCRIPTION_AND_QUEUE = (
+  addressSpaceName: string,
+  namespace: string,
+  type: string
+) => {
+  let filter = "";
+  if (addressSpaceName && addressSpaceName.trim() !== "") {
+    filter += "`$.metadata.name` LIKE '" + addressSpaceName + ".%' AND";
+  }
+  if (namespace && namespace.trim() !== "") {
+    filter += "`$.metadata.namespace` = '" + namespace + "'";
+  }
+  if (
+    type.trim().toLowerCase() === "subscription" ||
+    type.trim().toLowerCase() === "queue"
+  ) {
+    filter += " AND `$.spec.type` = 'deadletter'";
+  }
+  const ALL_DLQ_FOR_ADDRESS_SPACE = gql`
+    query all_addresses_for_addressspace_view {
+      addresses(
+        filter:"${filter}"
+      ) {
+        total
+        addresses {
+          metadata {
+            namespace
+            name
+          }
+          spec {
+            address
+            type
+            plan {
+              spec {
+                displayName
+              }
+            }
+          }
+        }
+      }
+    }
+  `;
+  return ALL_DLQ_FOR_ADDRESS_SPACE;
+};
 const RETURN_TOPIC_ADDRESSES_FOR_SUBSCRIPTION = (
   name: string,
   namespace: string,
@@ -628,6 +677,7 @@ const RETURN_ALL_ADDRESS_NAMES_OF_ADDRESS_SPACES_FOR_TYPEAHEAD_SEARCH = (
 export {
   DELETE_ADDRESS,
   PURGE_ADDRESS,
+  RETURN_DLQ_ADDRESSES_FOR_SUBSCRIPTION_AND_QUEUE,
   RETURN_ALL_ADDRESS_FOR_ADDRESS_SPACE,
   CURRENT_ADDRESS_SPACE_PLAN,
   RETURN_ADDRESS_DETAIL,